Switched the benchmarking stuff to more accurate nano-second stuff

+++ benchmarks/header/parseXHeadersBench.c
@@ -12,7 +12,7 @@ void parseXHeadersBench(int count) {
   buffer[length - 2] = '\n';
   buffer[length - 1] = '\0';
 
-  double total = parseHeadersBench(buffer);
+  long int total = parseHeadersBench(buffer);
 
   char nameBuffer[21];
   sprintf(nameBuffer, "Parse %06d-Headers", count);
