+++ benchmarks/queryparams/parseXQueryParams.c
@@ -19,7 +19,7 @@ void parseXQueryParamsBench(int count) {
   }
   buffer[length - 1] = '\0';
 
-  double total = parseQueryParamsBench(buffer);
+  long int total = parseQueryParamsBench(buffer);
 
   char nameBuffer[24];
   sprintf(nameBuffer, "Parse %06d-Parameters", count);
